replace double with float, have to change some test cases since the precision changes

+++ src/Algorithms/HypoTestPlaceHolder.h
@@ -27,7 +27,7 @@ public:
     HypoTestPlaceHolder();
     HypoTestPlaceHolder(const unordered_map<string, string>&);
 
-    double getProgress();
+    float getProgress();
     bool getIsRunning();
     void stop();
